+++ src/components/RadioPage/Desktop/styles/DesktopRadio.css
@@ -12,13 +12,6 @@
     box-shadow: 2px 7px 10px rgba(0,0,0,0.5);
 }
 
-.andrewRadioFace {
-    width: 200px;
-    height: auto;
-    margin: 0 auto;
-    margin-top: 10vh;
-}
-
 .footerText {
     color: white;
     font-size: 20px;
