+++ src/components/RadioPage/Mobile/styles/MobileRadio.css
@@ -1,0 +1,16 @@
+.andrewRadioFace {
+    width: 200px;
+    height: auto;
+    margin: 0 auto;
+    margin-top: 10vh;
+}
+
+.footerText {
+    color: white;
+    font-size: 20px;
+    text-align: center;
+    position: fixed;
+    left: 50%;
+    transform: translateX(-50%);
+    bottom: 20px;
+}
