+++ src/components/RadioPage/RadioPage.jsx
@@ -5,16 +5,13 @@ import DesktopRadio from './Desktop/DesktopRadio'
 class RadioPage extends React.Component {
     constructor(props) {
         super(props);
-        this.state = {
-            windowWidth: window.innerWidth
-        };
     }
 
     render() {
         // Set background to Blue
         document.body.style.backgroundColor = "#10C0FF";
 
-        if(this.state.windowWidth < 760) {
+        if(window.windowWidth < 760) {
             return <MobileRadio />
         } else {
             return <DesktopRadio />
